ios: silence some type conversion warnings

+++ ios/iosremote/PopoverView/PopoverView.m
@@ -825,7 +825,7 @@
 
 - (void)didTapButton:(UIButton *)sender
 {
-    int index = [subviewsArray indexOfObject:sender];
+    NSInteger index = [subviewsArray indexOfObject:sender];
     
     if (index == NSNotFound) {
         return;
