+++ ios/iosremote/iosremote/Communication/SlideShow.h
@@ -35,7 +35,7 @@
 - (void) putNotes: (NSString *)notes AtIndex: (uint) index;
 
 // Register an entry in the loadBuffer which loads the content (image or notes) directly into the desired UI element or wait until received the desired content from the server
-- (void) getContentAtIndex: (uint) index forView: (UIView*) view;
+- (void) getContentAtIndex: (NSUInteger) index forView: (UIView*) view;
 // Store the filename of the slideshow
 - (NSString *) title;
 - (void) setTitle:(NSString *)title;
